Fix NProgress not complete on page load
Issue:
When using Livewire with NProgress, the progress bar starts when navigating, but it does not complete when loading a new page.

Solution:
I fixed this by adding an event listener for livewire:navigated to properly trigger NProgress.done() once the page has finished loading. This ensures the progress bar completes and disappears as expected after navigation.